Introductory task infrastructure
Task contains description and date of creation (to track rollover in the future).

Added a TaskRepository interface to implement arbitrary kinds of repositories; currently uses an in-memory storage system but this can be changed later.

Added an adapter and fragment for the task list so that it can be displayed on the app.